tradovate-market-stream-main/requirements.txt
redis==5.0.1
loguru==0.7.2
typer==0.9.0
tqdm==4.66.1
rich==13.6.0
dotenv==0.9.9
websocket-client
requests==2.32.3
pip-system-certs
httpx
python-dateutil==2.9.0